+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - The Crafter can now only store 1 stack size pattern per slot (raoulvdberge)
 - You can now re-insert a Pattern in the pattern output slot in the Pattern Grid to modify an existing pattern (raoulvdberge)
 - Fixed not being able to use JEI R and U keys on Grid with tabs (raoulvdberge)
+- The Refined Storage jar is now signed (raoulvdberge)
 
 ### 1.5.21
 - Updated Portuguese (Brazilian) translation (Pinz714)